refactor: organize global `@/lib` files
- delete `colors.utils.ts` and `month-bullet.utils` from timeline's page file context
- remove unused `NumberToMonthEnum` and `numberToMonthEN`
- extract `monthsPTBR` to a const
- update imports too
